tasks/cars/src/main/java/cars/CarController.java
package cars;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
public class CarController {

    private CarService carService;

    public CarController(CarService carService) {
        this.carService = carService;
    }

    @GetMapping("/cars")
    public List<CarDto> list() {
        return carService.list();
    }

    @GetMapping("/types")
    public List<String> listType() {
        return carService.listType();
    }
}
